Move more methods to streaming's utils from the main file
